+++ src/compas_timber/connections/__init__.py
@@ -9,6 +9,9 @@ from .l_french_ridge_lap import LFrenchRidgeLapJoint
 from .lap_joint import LapJoint
 from .joint_candidate import JointCandidate
 from .joint_candidate import PlateJointCandidate
+from .topology_data import BeamTopologyData
+from .topology_data import PlateTopologyData
+from .topology_data import TopologyData
 from .solver import ConnectionSolver
 from .solver import PlateConnectionSolver
 from .solver import JointTopology
@@ -65,6 +68,9 @@ __all__ = [
     "LLapJoint",
     "JointCandidate",
     "PlateJointCandidate",
+    "BeamTopologyData",
+    "PlateTopologyData",
+    "TopologyData",
     "LFrenchRidgeLapJoint",
     "JointTopology",
     "ConnectionSolver",